+++ include/App.h
@@ -21,6 +21,6 @@ typedef struct {
 /*                          Functions                                          */
 
 void DataSentCallBack(const uint8_t *mac_addr, esp_now_send_status_t status);
-BOOL TimeStampHandler();
+BOOL MessageSendHandler(TimeStamp_t *dataPtr , MemoryConfig_t *memPtr);
