+++ src/SwTimer.cpp
@@ -1,0 +1,21 @@
+#include "SwTimer.h"
+
+
+
+/*
+void timerTask(void *pvParameters) {
+
+    while (1) {
+        vTaskDelay(MS_TO_TICKS(1));  // Increment timer every 1 ms
+        timer_ms++;
+    }
+}
+*/
+
+void TimerHandler()
+{
+
+
+
+
+}
